ref: refactoring add-item-form
